Move helper functions outside
